Improve diff feature
