+++ src/cmd/diff.js
@@ -1,9 +1,16 @@
+const log = require('../lib/log')
 const getPkg = require('../lib/get-pkg')
-const checkDiff = require('../lib/check-diff')
+const getDiff = require('../lib/get-diff')
+const logDiff = require('../lib/log-diff')
 let CWD = process.cwd()
 
 module.exports = async function diff () {
   const pkg = await getPkg()
   const { propertyId, experienceId } = pkg.meta
-  await checkDiff(CWD, propertyId, experienceId)
+  let diffs = await getDiff(CWD, propertyId, experienceId)
+  if (diffs.length) {
+    logDiff(diffs)
+  } else {
+    log.info('Both versions are the same!')
+  }
 }
